@@ -28,15 +28,15 @@ const (
 )
 
 // CacheEntry Every cache entry has this information
-type CacheEntry struct {
+type CacheEntry[T any] struct {
 	cacheKey              string     // key stringficated; needed for removal operation
 	lock                  sync.Mutex // lock for repeated requests
 	cond                  *sync.Cond // used in conjunction with the lock for repeated request until result is ready
-	postProcessedResponse interface{}
+	postProcessedResponse T
 	timestamp             time.Time // Last time accessed
 	expirationTime        time.Time
-	prev                  *CacheEntry
-	next                  *CacheEntry
+	prev                  *CacheEntry[T]
+	next                  *CacheEntry[T]
 	state                 int8 // AVAILABLE, COMPUTING, etc
 	err                   error
 }
@@ -46,45 +46,48 @@ type RequestError struct {
 	Code  int
 }
 
-type CacheDriver struct {
-	table             map[string]*CacheEntry
-	missCount         int
-	hitCount          int
-	ttl               time.Duration
-	head              CacheEntry // sentinel header node
-	lock              sync.Mutex
-	capacity          int
-	extendedCapacity  int
-	numEntries        int
-	toCompress        bool
-	toMapKey          func(key interface{}) (string, error)
-	valueToBytes      func(value interface{}) ([]byte, error)
-	bytesToValue      func([]byte) (interface{}, error)
-	preProcessRequest func(request interface{}, other ...interface{}) (interface{}, *RequestError)
-	callUServices     func(request, payload interface{}, other ...interface{}) (interface{}, *RequestError)
+// CacheDriver The cache itself T represents the request type and K the response type
+type CacheDriver[T any, K any] struct {
+	table            map[string]*CacheEntry[K]
+	missCount        int
+	hitCount         int
+	ttl              time.Duration
+	head             CacheEntry[K] // sentinel header node
+	lock             sync.Mutex
+	capacity         int
+	extendedCapacity int
+	numEntries       int
+	toCompress       bool
+	proccessor       ProccessorI[T, K]
+	compressor       CompressorI[K]
+	// toMapKey          func(key interface{}) (string, error)
+	// valueToBytes      func(value interface{}) ([]byte, error)
+	// bytesToValue      func([]byte) (interface{}, error)
+	// preProcessRequest func(request interface{}, other ...interface{}) (interface{}, *RequestError)
+	// callUServices     func(request, payload interface{}, other ...interface{}) (interface{}, *RequestError)
 }
 
-func (cache *CacheDriver) MissCount() int {
+func (cache *CacheDriver[T, K]) MissCount() int {
 	return cache.missCount
 }
 
-func (cache *CacheDriver) HitCount() int {
+func (cache *CacheDriver[T, K]) HitCount() int {
 	return cache.hitCount
 }
 
-func (cache *CacheDriver) Ttl() time.Duration {
+func (cache *CacheDriver[T, K]) Ttl() time.Duration {
 	return cache.ttl
 }
 
-func (cache *CacheDriver) Capacity() int {
+func (cache *CacheDriver[T, K]) Capacity() int {
 	return cache.capacity
 }
 
-func (cache *CacheDriver) ExtendedCapacity() int {
+func (cache *CacheDriver[T, K]) ExtendedCapacity() int {
 	return cache.extendedCapacity
 }
 
-func (cache *CacheDriver) NumEntries() int {
+func (cache *CacheDriver[T, K]) NumEntries() int {
 	return cache.numEntries
 }
 
@@ -103,12 +106,13 @@ func (cache *CacheDriver) NumEntries() int {
 //
 // callUService: is responsible for calling to the service and building a byte sequence corresponding to the
 // service response
-//
-func New(capacity int, capFactor float64, ttl time.Duration,
-	toMapKey func(key interface{}) (string, error),
-	preProcessRequest func(request interface{}, other ...interface{}) (interface{}, *RequestError),
-	callUServices func(request, payload interface{}, other ...interface{}) (interface{}, *RequestError),
-) *CacheDriver {
+func New[T any, K any](
+	capacity int,
+	capFactor float64,
+	ttl time.Duration,
+	proccessor ProccessorI[T, K],
+
+) *CacheDriver[T, K] {
 
 	if capFactor < 0.1 || capFactor > 3.0 {
 		panic(fmt.Sprintf("invalid capFactor %f. It should be in [0.1, 3]",
@@ -116,17 +120,18 @@ func New(capacity int, capFactor float64, ttl time.Duration,
 	}
 
 	extendedCapacity := math.Ceil((1.0 + capFactor) * float64(capacity))
-	ret := &CacheDriver{
-		missCount:         0,
-		hitCount:          0,
-		capacity:          capacity,
-		extendedCapacity:  int(extendedCapacity),
-		numEntries:        0,
-		ttl:               ttl,
-		table:             make(map[string]*CacheEntry, int(extendedCapacity)),
-		toMapKey:          toMapKey,
-		preProcessRequest: preProcessRequest,
-		callUServices:     callUServices,
+	ret := &CacheDriver[T, K]{
+		missCount:        0,
+		hitCount:         0,
+		capacity:         capacity,
+		extendedCapacity: int(extendedCapacity),
+		numEntries:       0,
+		ttl:              ttl,
+		table:            make(map[string]*CacheEntry[K], int(extendedCapacity)),
+		proccessor:       proccessor,
+		// toMapKey:          toMapKey,
+		// preProcessRequest: preProcessRequest,
+		// callUServices:     callUServices,
 	}
 	ret.head.prev = &ret.head
 	ret.head.next = &ret.head
@@ -164,28 +169,25 @@ func New(capacity int, capFactor float64, ttl time.Duration,
 //
 // callUService: is responsible for calling to the service and building a byte sequence corresponding to the
 // service response
-//
-func NewWithCompression(capacity int, capFactor float64, ttl time.Duration,
-	toMapKey func(key interface{}) (string, error),
-	valueToBytes func(value interface{}) ([]byte, error),
-	bytesToValue func([]byte) (interface{}, error),
-	preProcessRequest func(request interface{}, other ...interface{}) (interface{}, *RequestError),
-	callUServices func(request, payload interface{},
-		other ...interface{}) (interface{}, *RequestError),
-) (cache *CacheDriver) {
+func NewWithCompression[T any, K any](
+	capacity int,
+	capFactor float64,
+	ttl time.Duration,
+	proccessor ProccessorI[T, K],
+	compressor CompressorI[K],
+) (cache *CacheDriver[T, K]) {
 
-	cache = New(capacity, capFactor, ttl, toMapKey, preProcessRequest, callUServices)
+	cache = New(capacity, capFactor, ttl, proccessor)
 	if cache != nil {
 		cache.toCompress = true
-		cache.valueToBytes = valueToBytes
-		cache.bytesToValue = bytesToValue
+		cache.compressor = compressor
 	}
 
 	return cache
 }
 
 // Insert entry as the first item of cache (mru)
-func (cache *CacheDriver) insertAsMru(entry *CacheEntry) {
+func (cache *CacheDriver[T, K]) insertAsMru(entry *CacheEntry[K]) {
 	entry.prev = &cache.head
 	entry.next = cache.head.next
 	cache.head.next.prev = entry
@@ -193,18 +195,18 @@ func (cache *CacheDriver) insertAsMru(entry *CacheEntry) {
 }
 
 // Auto deletion of lru queue
-func (entry *CacheEntry) selfDeleteFromLRUList() {
+func (entry *CacheEntry[T]) selfDeleteFromLRUList() {
 	entry.prev.next = entry.next
 	entry.next.prev = entry.prev
 }
 
-func (cache *CacheDriver) becomeMru(entry *CacheEntry) {
+func (cache *CacheDriver[T, K]) becomeMru(entry *CacheEntry[K]) {
 	entry.selfDeleteFromLRUList()
 	cache.insertAsMru(entry)
 }
 
 // Rewove the last item in the list (lru); mutex must be taken. The entry becomes AVAILABLE
-func (cache *CacheDriver) evictLruEntry() (*CacheEntry, error) {
+func (cache *CacheDriver[T, K]) evictLruEntry() (*CacheEntry[K], error) {
 	entry := cache.head.prev // <-- LRU entry
 	if entry.state == COMPUTING {
 		err := errors.New("LRU entry is in COMPUTING state. This could be a bug or a cache misconfiguration")
@@ -216,8 +218,9 @@ func (cache *CacheDriver) evictLruEntry() (*CacheEntry, error) {
 	return entry, nil
 }
 
-func (cache *CacheDriver) allocateEntry(cacheKey string,
-	currTime time.Time) (entry *CacheEntry, err error) {
+func (cache *CacheDriver[T, K]) allocateEntry(
+	cacheKey string,
+	currTime time.Time) (entry *CacheEntry[K], err error) {
 
 	if cache.numEntries == cache.capacity {
 		entry, err = cache.evictLruEntry()
@@ -225,7 +228,7 @@ func (cache *CacheDriver) allocateEntry(cacheKey string,
 			return nil, err
 		}
 	} else {
-		entry = new(CacheEntry)
+		entry = new(CacheEntry[K])
 		entry.cond = sync.NewCond(&entry.lock)
 		cache.numEntries++
 	}
@@ -234,7 +237,8 @@ func (cache *CacheDriver) allocateEntry(cacheKey string,
 	entry.state = AVAILABLE
 	entry.timestamp = currTime
 	entry.expirationTime = currTime.Add(cache.ttl)
-	entry.postProcessedResponse = nil // should dispose any allocated result
+	var zeroK K
+	entry.postProcessedResponse = zeroK // should dispose any allocated result
 	cache.table[cacheKey] = entry
 	return entry, nil
 }
@@ -243,30 +247,21 @@ func (cache *CacheDriver) allocateEntry(cacheKey string,
 // immediately returns the cached entry. If the request is the first, then it blocks until the result is
 // ready. If the request is not the first but the result is not still ready, then it blocks
 // until the result is ready
-func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
-	other ...interface{}) (interface{}, *RequestError) {
+func (cache *CacheDriver[T, K]) RetrieveFromCacheOrCompute(request T) (K, *RequestError) {
 
 	var requestError *RequestError
+	var zeroK K
+	payload := request
 
-	var payload interface{}
-	if cache.preProcessRequest != nil {
-		payload, requestError = cache.preProcessRequest(request, other...)
-		if requestError != nil {
-			return nil, requestError
-		}
-	} else {
-		payload = request
-	}
-
-	cacheKey, err := cache.toMapKey(payload)
+	cacheKey, err := cache.proccessor.ToMapKey(payload)
 	if err != nil {
-		return nil, &RequestError{
+		return zeroK, &RequestError{
 			Error: err,
 			Code:  Status4xx,
 		}
 	}
 
-	var entry *CacheEntry
+	var entry *CacheEntry[K]
 	var hit bool
 	currTime := time.Now()
 	cache.lock.Lock()
@@ -285,12 +280,12 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 		}
 		defer entry.lock.Unlock()
 		if entry.state == FAILED5xx {
-			return nil, &RequestError{
+			return zeroK, &RequestError{
 				Error: entry.err,
 				Code:  Status5xxCached, // include 4xx and 5xx
 			}
 		} else if entry.state == FAILED4xx {
-			return nil, &RequestError{
+			return zeroK, &RequestError{
 				Error: entry.err,
 				Code:  Status4xxCached, // include 4xx and 5xx
 			}
@@ -299,16 +294,25 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 		entry.expirationTime = currTime.Add(cache.ttl)
 
 		if withCompression {
-			buf, err := lz4Decompress(entry.postProcessedResponse.([]byte))
+
+			data, err := cache.compressor.ValueToBytes(entry.postProcessedResponse)
 			if err != nil {
-				return nil, &RequestError{
+				return zeroK, &RequestError{
+					Error: errors.New("cannot convert stored message"),
+					Code:  Status5xx, // include 4xx and 5xx
+				}
+			}
+			buf, err := lz4Decompress(data)
+			if err != nil {
+				return zeroK, &RequestError{
 					Error: errors.New("cannot decompress stored message"),
 					Code:  Status5xx, // include 4xx and 5xx
 				}
 			}
-			result, err := cache.bytesToValue(buf)
+			// result, err := cache.bytesToValue(buf)
+			result, err := cache.compressor.BytesToValue(buf)
 			if err != nil {
-				return nil, &RequestError{
+				return zeroK, &RequestError{
 					Error: errors.New("cannot convert decompressed stored message"),
 					Code:  Status5xx, // include 4xx and 5xx
 				}
@@ -324,7 +328,8 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 	entry, err = cache.allocateEntry(cacheKey, currTime)
 	if err != nil {
 		cache.lock.Unlock() // an error getting cache entry ==> we invoke directly the uservice
-		return cache.callUServices(request, payload, other...)
+		// return cache.callUServices(request, payload, other...)
+		return cache.proccessor.CallUServices(request)
 	}
 
 	entry.state = COMPUTING
@@ -335,8 +340,8 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 	entry.lock.Lock() // other requests will wait for until postProcessedResponse is gotten
 	defer entry.lock.Unlock()
 
-	var retVal interface{} = nil // Explicit initialization for understanding flow!
-	retVal, requestError = cache.callUServices(request, payload, other...)
+	// retVal, requestError = cache.callUServices(request, payload, other...)
+	retVal, requestError := cache.proccessor.CallUServices(request)
 	if requestError != nil {
 		switch {
 		case requestError.Code == Status4xx || requestError.Code == Status4xxCached:
@@ -352,7 +357,8 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 	}
 
 	if withCompression {
-		buf, err := cache.valueToBytes(retVal) // transforms retVal into a []byte ready for compression
+		// buf, err := cache.valueToBytes(retVal) // transforms retVal into a []byte ready for compression
+		buf, err := cache.compressor.ValueToBytes(retVal) // transforms retVal into a []byte ready for compression
 		if err != nil {
 			entry.state = FAILED5xx
 		}
@@ -360,7 +366,13 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 		if err != nil {
 			entry.state = FAILED5xx
 		}
-		entry.postProcessedResponse = lz4Buf // stores the compressed representation
+		// if you want to store the compressed representation
+		// your type should be a []byte, to validate it we use the interface{}(lz4Buf).(K)
+		if resp, ok := interface{}(lz4Buf).(K); !ok {
+			entry.state = FAILED5xx
+		} else {
+			entry.postProcessedResponse = resp // stores the compressed representation
+		}
 	} else {
 		entry.postProcessedResponse = retVal
 	}
@@ -371,7 +383,7 @@ func (cache *CacheDriver) RetrieveFromCacheOrCompute(request interface{},
 }
 
 // remove entry from cache.Mutex must be taken
-func (cache *CacheDriver) remove(entry *CacheEntry) {
+func (cache *CacheDriver[T, K]) remove(entry *CacheEntry[K]) {
 	entry.selfDeleteFromLRUList()
 	cache.table[entry.cacheKey] = nil
 	delete(cache.table, entry.cacheKey)
@@ -379,8 +391,9 @@ func (cache *CacheDriver) remove(entry *CacheEntry) {
 }
 
 // has return true is state in the cache
-func (cache *CacheDriver) has(val interface{}) bool {
-	key, err := cache.toMapKey(val)
+func (cache *CacheDriver[T, K]) has(val T) bool {
+	key, err := cache.proccessor.ToMapKey(val)
+	// key, err := cache.toMapKey(val)
 	if err != nil {
 		return false
 	}
@@ -389,7 +402,7 @@ func (cache *CacheDriver) has(val interface{}) bool {
 }
 
 // Return the lru without moving it from the queue
-func (cache *CacheDriver) getLru() *CacheEntry {
+func (cache *CacheDriver[T, K]) getLru() *CacheEntry[K] {
 	if cache.numEntries == 0 {
 		return nil
 	}
@@ -397,7 +410,7 @@ func (cache *CacheDriver) getLru() *CacheEntry {
 }
 
 // Return the mru without moving it from the queue
-func (cache *CacheDriver) getMru() *CacheEntry {
+func (cache *CacheDriver[T, K]) getMru() *CacheEntry[K] {
 	if cache.numEntries == 0 {
 		return nil
 	}
@@ -405,24 +418,24 @@ func (cache *CacheDriver) getMru() *CacheEntry {
 }
 
 // CacheIt Iterator on cache entries. Go from MUR to LRU
-type CacheIt struct {
-	cachePtr *CacheDriver
-	curr     *CacheEntry
+type CacheIt[T any, K any] struct {
+	cachePtr *CacheDriver[T, K]
+	curr     *CacheEntry[K]
 }
 
-func (cache *CacheDriver) NewCacheIt() *CacheIt {
-	return &CacheIt{cachePtr: cache, curr: cache.head.next}
+func (cache *CacheDriver[T, K]) NewCacheIt() *CacheIt[T, K] {
+	return &CacheIt[T, K]{cachePtr: cache, curr: cache.head.next}
 }
 
-func (it *CacheIt) HasCurr() bool {
+func (it *CacheIt[T, K]) HasCurr() bool {
 	return it.curr != &it.cachePtr.head
 }
 
-func (it *CacheIt) GetCurr() *CacheEntry {
+func (it *CacheIt[T, K]) GetCurr() *CacheEntry[K] {
 	return it.curr
 }
 
-func (it *CacheIt) Next() *CacheEntry {
+func (it *CacheIt[T, K]) Next() *CacheEntry[K] {
 	if !it.HasCurr() {
 		return nil
 	}
@@ -439,7 +452,7 @@ type CacheState struct {
 }
 
 // GetState Return a json containing the cache state. Use the internal mutex. Be careful with a deadlock
-func (cache *CacheDriver) GetState() (string, error) {
+func (cache *CacheDriver[T, K]) GetState() (string, error) {
 
 	cache.lock.Lock()
 	defer cache.lock.Unlock()
@@ -460,7 +473,7 @@ func (cache *CacheDriver) GetState() (string, error) {
 }
 
 // helper that does not take lock
-func (cache *CacheDriver) clean() error {
+func (cache *CacheDriver[T, K]) clean() error {
 
 	for it := cache.NewCacheIt(); it.HasCurr(); it.Next() {
 		entry := it.GetCurr()
@@ -486,8 +499,7 @@ func (cache *CacheDriver) clean() error {
 // state.
 //
 // Uses internal lock
-//
-func (cache *CacheDriver) Clean() error {
+func (cache *CacheDriver[T, K]) Clean() error {
 
 	cache.lock.Lock()
 	defer cache.lock.Unlock()
@@ -495,7 +507,7 @@ func (cache *CacheDriver) Clean() error {
 	return cache.clean()
 }
 
-func (cache *CacheDriver) Set(capacity int, ttl time.Duration) error {
+func (cache *CacheDriver[T, K]) Set(capacity int, ttl time.Duration) error {
 
 	cache.lock.Lock()
 	defer cache.lock.Unlock()
